moved authContext and useAuth into seperate files in attempt to resolve hmr issue
