+++ src/client/components/auth/PrivateRoute.jsx
@@ -1,4 +1,5 @@
-import { useAuth } from './AuthProvider';
+//import { useAuth } from './AuthProvider';
+import { useAuth } from './useAuth';
 import { Navigate } from 'react-router-dom';
 //PrivateRoute is a wrapper for routes
 //redirects to login if user is not authenticated
